MariaDB: use Galera replication
The more I stare at the MariaDB non-Galera replication docs, the more
I'm convinced that it's not meant to be used. They seem to be under the
impression that promoting a new leader should *only* be done by graceful
failover; I literally can't figure out how to handle an actual node
crash.