Try an alternate approach to bootstrapping
Instead of bootstrapping a node with an empty cluster then trying to
restart it later, we put the bootstrap state into a file and check for
its existence on launch. We also use the hostname to figure out whether
or not to bootstrap, which means we can use a single image.